backend/requirements.txt
# API
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Image
Pillow>=10.0.0

# Providers (optional: install only what you use)
openai>=1.12.0
google-generativeai>=0.4.0

# Local model (heavy: only if using provider=local)
# torch>=2.0.0
# transformers>=4.36.0
# accelerate>=0.25.0
